# lib/api-zod/src/generated/types/reconciliationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FreightFlow - Freight Invoice Automation & Reconciliation API
 * OpenAPI spec version: 0.1.0
 */

export interface ReconciliationRequest {
  invoiceIds: number[];
}