# lib/api-zod/src/generated/types/shipment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FreightFlow - Freight Invoice Automation & Reconciliation API
 * OpenAPI spec version: 0.1.0
 */
import type { ShipmentStatus } from "./shipmentStatus";

export interface Shipment {
  id: number;
  shipmentNumber: string;
  vendorId: number;
  vendorName: string;
  origin: string;
  destination: string;
  status: ShipmentStatus;
  scheduledDelivery?: Date | null;
  actualDelivery?: Date | null;
  agreedFreightCost: number;
  weightKg?: number | null;
  description?: string | null;
  createdAt: Date;
}