# lib/api-zod/src/generated/types/updateInvoiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FreightFlow - Freight Invoice Automation & Reconciliation API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateInvoiceRequestStatus } from "./updateInvoiceRequestStatus";

export interface UpdateInvoiceRequest {
  status?: UpdateInvoiceRequestStatus;
  shipmentId?: number | null;
  discrepancyNotes?: string | null;
  amount?: number;
  fuelSurcharge?: number | null;
  otherCharges?: number | null;
  gstAmount?: number | null;
}