# lib/api-zod/src/generated/types/updateShipmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FreightFlow - Freight Invoice Automation & Reconciliation API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateShipmentRequestStatus } from "./updateShipmentRequestStatus";

export interface UpdateShipmentRequest {
  status?: UpdateShipmentRequestStatus;
  actualDelivery?: Date | null;
  agreedFreightCost?: number;
}